+++ notification-center/test/Dignite.Abp.NotificationCenter.TestBase/NotificationStore_Tests.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Dignite.Abp.Notifications;
@@ -236,6 +237,88 @@ public abstract class NotificationStore_Tests<TStartupModule> : NotificationCent
             (await store.IsSubscribedAsync(bothScopes, "order.shipped", "Demo.Order", "42")).ShouldBeFalse();
             (await store.GetSubscriptionsAsync(bothScopes)).ShouldHaveSingleItem()
                 .EntityTypeName.ShouldBeNull();
+        });
+    }
+
+    [Fact]
+    public async Task Batched_subscription_pages_are_globally_distinct_stable_and_exact()
+    {
+        var users = Enumerable.Range(0, 5)
+            .Select(_ => Guid.NewGuid())
+            .OrderBy(userId => userId)
+            .ToArray();
+
+        await WithUnitOfWorkAsync(async () =>
+        {
+            var store = GetRequiredService<INotificationStore>();
+            await store.InsertSubscriptionAsync(NewSubscription(users[0]));
+            await store.InsertSubscriptionAsync(NewSubscription(users[1]));
+            await store.InsertSubscriptionAsync(NewSubscription(users[1], "Demo.Order", "42"));
+            await store.InsertSubscriptionAsync(NewSubscription(users[2], "Demo.Order", "42"));
+            await store.InsertSubscriptionAsync(NewSubscription(users[3], "Demo.Order", "42"));
+            await store.InsertSubscriptionAsync(NewSubscription(users[4], "Demo.Order", "99"));
+        });
+
+        await WithUnitOfWorkAsync(async () =>
+        {
+            var store = (IBatchedNotificationStore)GetRequiredService<INotificationStore>();
+            var firstPage = await store.GetSubscriptionUserIdsAsync(
+                "order.shipped", "Demo.Order", "42", null, 2);
+            var secondPage = await store.GetSubscriptionUserIdsAsync(
+                "order.shipped", "Demo.Order", "42", firstPage[^1], 2);
+            var exactBoundaryPage = await store.GetSubscriptionUserIdsAsync(
+                "order.shipped", "Demo.Order", "42", secondPage[^1], 2);
+
+            firstPage.ShouldBe(users.Take(2));
+            secondPage.ShouldBe(users.Skip(2).Take(2));
+            exactBoundaryPage.ShouldBeEmpty();
+            firstPage.Concat(secondPage).Distinct().Count().ShouldBe(4);
+        });
+    }
+
+    [Fact]
+    public async Task Subscription_keyset_does_not_repeat_or_skip_when_rows_before_the_cursor_change()
+    {
+        var a = Guid.Parse("10000000-0000-0000-0000-000000000000");
+        var insertedBeforeCursor = Guid.Parse("20000000-0000-0000-0000-000000000000");
+        var b = Guid.Parse("30000000-0000-0000-0000-000000000000");
+        var c = Guid.Parse("50000000-0000-0000-0000-000000000000");
+        var insertedAfterCursor = Guid.Parse("60000000-0000-0000-0000-000000000000");
+        var d = Guid.Parse("70000000-0000-0000-0000-000000000000");
+
+        await WithUnitOfWorkAsync(async () =>
+        {
+            var store = GetRequiredService<INotificationStore>();
+            foreach (var userId in new[] { a, b, c, d })
+            {
+                await store.InsertSubscriptionAsync(NewSubscription(userId));
+            }
+        });
+
+        List<Guid>? firstPage = null;
+        await WithUnitOfWorkAsync(async () =>
+        {
+            firstPage = await ((IBatchedNotificationStore)GetRequiredService<INotificationStore>())
+                .GetSubscriptionUserIdsAsync("order.shipped", null, null, null, 2);
+        });
+        firstPage.ShouldBe(new[] { a, b });
+
+        await WithUnitOfWorkAsync(async () =>
+        {
+            var store = GetRequiredService<INotificationStore>();
+            await store.DeleteSubscriptionAsync(a, "order.shipped", null, null);
+            await store.InsertSubscriptionAsync(NewSubscription(insertedBeforeCursor));
+            await store.InsertSubscriptionAsync(NewSubscription(insertedAfterCursor));
+        });
+
+        await WithUnitOfWorkAsync(async () =>
+        {
+            var secondPage = await ((IBatchedNotificationStore)GetRequiredService<INotificationStore>())
+                .GetSubscriptionUserIdsAsync("order.shipped", null, null, firstPage![^1], 10);
+
+            secondPage.ShouldBe(new[] { c, insertedAfterCursor, d });
+            secondPage.ShouldNotContain(b);
+            secondPage.ShouldNotContain(insertedBeforeCursor);
         });
     }